Updated css styling on modal cards in account page and on todolist text

+++ amma-track-app/src/pages/Account/Account.jsx
@@ -13,7 +13,15 @@ import "./Account.css";
 import cameraIcon from "../../images/camera-icon.svg";
 import { validatePassword } from "../../validations/validate";
 import { db } from "../../config/firebaseConfig";
-import { collection, updateDoc, getDocs, deleteDoc, doc } from "firebase/firestore";
+import {
+  collection,
+  updateDoc,
+  getDocs,
+  deleteDoc,
+  doc
+} from "firebase/firestore";
+import DeleteModal from "../../components/CardModal/DeleteModal";
+import { Button } from "../../components/Button/Button";
 
 export default function Account() {
   const dispatch = useDispatch();
@@ -34,16 +42,18 @@ export default function Account() {
   }
 
   const changeAccountData = async (data) => {
-    const snapshot = await getDocs(usersCollection)
-    for(let userDoc of snapshot.docs.filter(doc => doc.data().id === user.id)) {
-        if (userDoc.id) {
-            await updateDoc(doc(db, "users", userDoc.id), {
-              ...userDoc.data(),
-              ...data,
-            });
-        } 
+    const snapshot = await getDocs(usersCollection);
+    for (let userDoc of snapshot.docs.filter(
+      (doc) => doc.data().id === user.id
+    )) {
+      if (userDoc.id) {
+        await updateDoc(doc(db, "users", userDoc.id), {
+          ...userDoc.data(),
+          ...data
+        });
+      }
     }
-  }
+  };
 
   const handlePasswordChange = async (e) => {
     e.preventDefault();
@@ -56,7 +66,7 @@ export default function Account() {
       setConfirmPassword("");
       setFeedbackMessage("Password successfully changed.");
       setShowChangePasswordModal(false);
-      await changeAccountData({password: newPassword});
+      await changeAccountData({ password: newPassword });
     } else if (newPassword !== confirmPassword) {
       setPasswordErrors({
         ...errors,
@@ -76,8 +86,8 @@ export default function Account() {
     window.localStorage.setItem("loggedUser", JSON.stringify(updatedUser));
     setShowChangeUsernameModal(false);
     setFeedbackMessage("Username successfully changed.");
-    const snapshot = await getDocs(usersCollection)
-    await changeAccountData({userName: newUsername});
+    const snapshot = await getDocs(usersCollection);
+    await changeAccountData({ userName: newUsername });
   };
 
   const handleSwitchAccount = () => {
@@ -94,10 +104,11 @@ export default function Account() {
     dispatch(deleteAccount({ userId: user.id }));
     dispatch(logOut());
     setShowDeleteModal(false);
-    const snapshot = await getDocs(usersCollection)
-    for (let userDoc of snapshot.docs.filter(doc => doc.data().id === user.id)) {
-      if (userDoc.id)
-        await deleteDoc(doc(db, "users", userDoc.id))
+    const snapshot = await getDocs(usersCollection);
+    for (let userDoc of snapshot.docs.filter(
+      (doc) => doc.data().id === user.id
+    )) {
+      if (userDoc.id) await deleteDoc(doc(db, "users", userDoc.id));
     }
     navigate("/");
   };
@@ -110,8 +121,8 @@ export default function Account() {
         dispatch(updateUserImage({ userId: user.id, imageUrl: reader.result }));
       };
       reader.readAsDataURL(file);
-      console.log(file)
-      await changeAccountData({image: file.name});
+      console.log(file);
+      await changeAccountData({ image: file.name });
     }
   };
